backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.1.8
codecov==2.1.13
colorama==0.4.6
coverage==7.12.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fpl==0.6.35
frozenlist==1.8.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.4.1
psycopg2-binary==2.9.10
PTable==0.9.2
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==8.3.4
pytest-aiohttp==1.1.0
pytest-asyncio==1.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
stripe==13.0.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0
